requirements.txt
asgiref==3.7.2
Django==4.2.8
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
phonenumbers==8.13.26
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
